+++ mission/chapter08/src/pages/auth/AuthPage.tsx
@@ -1,91 +1,17 @@
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useMutation } from '@tanstack/react-query';
-import { useForm } from 'react-hook-form';
-import { Link, useLocation, useNavigate } from 'react-router';
-import AuthInput from '../../components/auth/AuthInput';
-import { signin, signup } from '../../apis/authApi';
-import { API_BASE_URL } from '../../apis/http';
-import { API_AUTH_PATHS, ROUTES } from '../../constants/paths';
-import {
-  signinFormSchema,
-  signupFormSchema,
-  type SigninFormValues,
-  type SignupFormValues,
-} from '../../schemas/authFormSchema';
-import { useAuth } from '../../contexts/AuthContext';
+import { Link, useLocation } from 'react-router';
+import SigninForm from '../../components/auth/SigninForm';
+import SignupForm from '../../components/auth/SignupForm';
+import { ROUTES } from '../../constants/paths';
 
 function AuthPage() {
-  const navigate = useNavigate();
   const location = useLocation();
-  const { login } = useAuth();
   const isSigninPage = location.pathname === ROUTES.authSignin;
   const from = (location.state as { from?: string } | null)?.from ?? ROUTES.home;
-
-  const signinForm = useForm<SigninFormValues>({
-    resolver: zodResolver(signinFormSchema),
-    mode: 'onTouched',
-    defaultValues: { email: '', password: '' },
-  });
-
-  const signupForm = useForm<SignupFormValues>({
-    resolver: zodResolver(signupFormSchema),
-    mode: 'onTouched',
-    defaultValues: { name: '', email: '', password: '', bio: '', avatar: '' },
-  });
-
-  // ── 로그인 Mutation ──────────────────────────────────────
-  const signinMutation = useMutation({
-    mutationFn: (values: SigninFormValues) => signin(values),
-    onSuccess: (result) => {
-      if (!result.data) throw new Error('로그인 응답 데이터가 없습니다.');
-      login(result.data.accessToken, result.data.refreshToken, result.data.name);
-      navigate(from, { replace: true });
-    },
-  });
-
-  // ── 회원가입 Mutation ─────────────────────────────────────
-  const signupMutation = useMutation({
-    mutationFn: (values: SignupFormValues) => {
-      const payload = {
-        name: values.name,
-        email: values.email,
-        password: values.password,
-        ...(values.bio?.trim() ? { bio: values.bio.trim() } : {}),
-        ...(values.avatar?.trim() ? { avatar: values.avatar.trim() } : {}),
-      };
-      return signup(payload);
-    },
-    onSuccess: () => {
-      navigate(ROUTES.authSignin, { replace: true });
-    },
-  });
-
-  const handleSigninSubmit = signinForm.handleSubmit((values) => {
-    signinMutation.mutate(values);
-  });
-
-  const handleSignupSubmit = signupForm.handleSubmit((values) => {
-    signupMutation.mutate(values);
-  });
-
-  const handleGoogleSignin = () => {
-    window.location.href = `${API_BASE_URL}${API_AUTH_PATHS.googleLogin}`;
-  };
-
-  const {
-    register: registerSignin,
-    formState: { errors: signinErrors },
-  } = signinForm;
-
-  const {
-    register: registerSignup,
-    formState: { errors: signupErrors },
-  } = signupForm;
 
   return (
     <div className="min-h-screen bg-[#111111] flex items-center justify-center px-4">
       <div className="w-full max-w-md">
-        <Link to="/" className="mb-8 block text-center text-2xl font-extrabold text-pink-500">
+        <Link to={ROUTES.home} className="mb-8 block text-center text-2xl font-extrabold text-pink-500">
           DOLIGO
         </Link>
 
@@ -97,110 +23,7 @@ function AuthPage() {
               : '간단한 정보 입력 후 바로 서비스를 시작할 수 있어요.'}
           </p>
 
-          {/* 로그인 에러 */}
-          {signinMutation.isError && (
-            <p className="mb-3 rounded-lg bg-red-500/10 px-3 py-2 text-sm text-red-400">
-              {signinMutation.error instanceof Error
-                ? signinMutation.error.message
-                : '로그인에 실패했습니다.'}
-            </p>
-          )}
-          {/* 회원가입 에러 */}
-          {signupMutation.isError && (
-            <p className="mb-3 rounded-lg bg-red-500/10 px-3 py-2 text-sm text-red-400">
-              {signupMutation.error instanceof Error
-                ? signupMutation.error.message
-                : '회원가입에 실패했습니다.'}
-            </p>
-          )}
-
-          {isSigninPage ? (
-            <form onSubmit={handleSigninSubmit} className="grid gap-3">
-              <AuthInput
-                id="signin-email"
-                label="이메일"
-                type="email"
-                placeholder="example@email.com"
-                autoComplete="email"
-                registration={registerSignin('email')}
-                error={signinErrors.email?.message}
-              />
-              <AuthInput
-                id="signin-password"
-                label="비밀번호"
-                type="password"
-                placeholder="비밀번호를 입력하세요"
-                autoComplete="current-password"
-                registration={registerSignin('password')}
-                error={signinErrors.password?.message}
-              />
-              <button
-                type="submit"
-                className="h-11 rounded-xl bg-pink-500 text-sm font-semibold text-white transition hover:bg-pink-600 disabled:cursor-not-allowed disabled:opacity-60"
-                disabled={signinMutation.isPending}
-              >
-                {signinMutation.isPending ? '로그인 중...' : '로그인'}
-              </button>
-              <button
-                type="button"
-                className="h-11 rounded-xl border border-white/20 bg-transparent text-sm font-semibold text-slate-300 transition hover:bg-white/10"
-                onClick={handleGoogleSignin}
-              >
-                Google로 로그인
-              </button>
-            </form>
-          ) : (
-            <form onSubmit={handleSignupSubmit} className="grid gap-3">
-              <AuthInput
-                id="signup-name"
-                label="이름"
-                placeholder="이름을 입력하세요"
-                autoComplete="name"
-                registration={registerSignup('name')}
-                error={signupErrors.name?.message}
-              />
-              <AuthInput
-                id="signup-email"
-                label="이메일"
-                type="email"
-                placeholder="example@email.com"
-                autoComplete="email"
-                registration={registerSignup('email')}
-                error={signupErrors.email?.message}
-              />
-              <AuthInput
-                id="signup-password"
-                label="비밀번호"
-                type="password"
-                placeholder="비밀번호를 입력하세요"
-                autoComplete="new-password"
-                registration={registerSignup('password')}
-                error={signupErrors.password?.message}
-              />
-              <AuthInput
-                id="signup-bio"
-                label="자기소개 (선택)"
-                placeholder="자기소개를 입력하세요"
-                registration={registerSignup('bio')}
-                error={signupErrors.bio?.message}
-              />
-              <AuthInput
-                id="signup-avatar"
-                label="아바타 URL (선택)"
-                type="url"
-                placeholder="https://..."
-                registration={registerSignup('avatar')}
-                error={signupErrors.avatar?.message}
-              />
-              <button
-                type="submit"
-                className="h-11 rounded-xl bg-pink-500 text-sm font-semibold text-white transition hover:bg-pink-600 disabled:cursor-not-allowed disabled:opacity-60"
-                disabled={signupMutation.isPending}
-              >
-                {signupMutation.isPending ? '회원가입 중...' : '회원가입'}
-              </button>
-            </form>
-          )}
+          {isSigninPage ? <SigninForm redirectTo={from} /> : <SignupForm />}
 
           <p className="mt-4 text-center text-sm text-slate-500">
             {isSigninPage ? (
